+++ stage2/driver/port.h
@@ -1,6 +1,8 @@
 #pragma once
 #include <stdint.h>
 
+#define PORT_UNUSED 0x80
+
 uint8_t port8_in(uint16_t id);
 void port8_out(uint16_t id, uint8_t value);
 
@@ -9,3 +11,5 @@ void port16_out(uint16_t id, uint16_t value);
 
 uint32_t port32_in(uint16_t id);
 void port32_out(uint16_t id, uint32_t value);
+
+void port_wait(void);