+++ stage2/global.h
@@ -1,13 +1,10 @@
 #pragma once
 
-#define __packed __attribute__((packed))
-#define __cdecl  __attribute__((cdecl))
-
-#ifndef _MSC_VER
-#	define __align(B) __attribute__((aligned(B)))
-#else //_MSC_VER
-#	define __align(B) __declspec(align(B))
-#endif//_MSC_VER
+#define __packed    __attribute__((packed))
+#define __cdecl     __attribute__((cdecl))
+#define __asmcall   __cdecl extern
+#define __align(X)  __attribute__((aligned(X)))
+#define __interrupt __attribute__((interrupt))
 
 #define MAX(A, B) ((A) > (B) ? (A) : (B))
 #define MIN(A, B) ((A) < (B) ? (A) : (B))