+++ stage2/interface/console.h
@@ -9,7 +9,6 @@
 #define CONSOLE_FOREGROUND(Color) ((Color) << 0x00)
 #define CONSOLE_BACKGROUND(Color) ((Color) << 0x04)
 #define CONSOLE_COLOR(Foreground, Background) (CONSOLE_FOREGROUND(Foreground) | CONSOLE_BACKGROUND(Background))
-#define CONSOLE_COLOR_DEFAULT CONSOLE_COLOR(CONSOLE_LIGHT_GRAY, CONSOLE_BLACK)
 
 #define CONSOLE_CURSOR(Row, Col) ((Col) + (Row) * CONSOLE_COLS)
 #define CONSOLE_CURSOR_BLOCK 0x0F
@@ -37,8 +36,10 @@ typedef enum {
 extern uint16_t console_cursor;
 extern uint8_t console_color;
 
+void putc(char chr);
+void puts(const char *str);
+
 void console_init(void);
-void console_print(char chr);
 void console_flush(void);
 void console_clear(void);
 void console_scroll(uint8_t rows);