+++ stage2/protocol/chainload.c
@@ -14,7 +14,7 @@ bool chainload_init(disk_t disk)
 	return true;
 }
 
-bool chainload_name(char buffer[11])
+bool chainload_name(char *buffer)
 {
 	if (*(uint8_t*)0x7C26 == 0x29) {
 		memcpy(buffer, (char*)0x7C2B, 11);